+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -15,11 +15,11 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Controller;
 import org.apache.coyote.ControllerMappings;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.request.HttpHeaders;
+import org.apache.coyote.http11.request.spec.HttpHeaders;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.ResourceLocator;
 import org.apache.coyote.http11.request.HttpRequestHandler;
-import org.apache.coyote.http11.request.StartLine;
+import org.apache.coyote.http11.request.spec.StartLine;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
